fix page router
