+++ src/views/account-profile/index.scss
@@ -1,11 +1,11 @@
-.account-settings-container {
+.account-profile-container {
   display: flex;
   padding: 20px;
   min-height: calc(100vh - 56px - 90px);
   box-sizing: border-box;
 }
 
-.account-settings-navbox {
+.account-profile-navbox {
   width: 234px;
   padding: 8px;
   background-color: #ffffff;
@@ -14,7 +14,7 @@
   flex-direction: column;
 }
 
-.account-settings-navitem {
+.account-profile-navitem {
   display: flex;
   height: 40px;
   border-radius: 4px;
@@ -24,16 +24,16 @@
   padding: 0 12px;
   margin-bottom: 8px;
 }
-.account-settings-navitem.active,.account-settings-navitem:hover {
+.account-profile-navitem.active,.account-profile-navitem:hover {
   background: #E2F3FE;
   color: #237FFA;
 }
-.account-settings-navitem .nav-icon {
+.account-profile-navitem .nav-icon {
   font-size: 16px;
   margin-right: 8px;
 }
 
-.account-settings-contentbox {
+.account-profile-contentbox {
   flex: 1;
   padding: 0 20px;
   background-color: #ffffff;
@@ -45,17 +45,18 @@
     color: #1F2733;
     box-shadow: inset 0px -1px 0px 0px #EBEDF0;
   }
-  &-formcontainer {
-    margin-top: var(--hi-v4-spacing-10, 20px);
-    min-width: 520px;
-    max-width: 720px;
-  }
 }
-.account-settings-avatar {
-  .hi-v4-upload__item {
-    margin: 0;
-    width: 116px;
-    height: 116px;
-    line-height: 114px;
+
+.account-setting {
+  margin-top: var(--hi-v4-spacing-10, 20px);
+  min-width: 520px;
+  max-width: 720px;
+  .account-settings-avatar {
+    .hi-v4-upload__item {
+      margin: 0;
+      width: 116px;
+      height: 116px;
+      line-height: 114px;
+    }
   }
 }
